moved files into /pages/[lng] director
